python/solution_tasks/simple_hr/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-hr"
version = "1.0.0"
description = "Simple HR Management System built with Flask"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Flask",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "Flask>=3.0.0",
    "Flask-SQLAlchemy>=3.0.0",
    "Flask-Login>=0.6.0",
    "Flask-WTF>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "isort>=5.13.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-flask>=1.3.0",
]

# ============================================================================
# Black - форматтер кода
# ============================================================================
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | venv
  | \.mypy_cache
  | \.pytest_cache
  | \.tox
  | build
  | dist
  | migrations
  | instance
)/
'''

# ============================================================================
# Ruff - быстрый линтер
# ============================================================================
[tool.ruff]
line-length = 100
target-version = "py39"
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "migrations",
    "instance",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "C",      # flake8-comprehensions
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
    "N",      # pep8-naming
    "S",      # flake8-bandit (security)
    "T20",    # flake8-print
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "ARG",    # flake8-unused-arguments
]

ignore = [
    "E501",   # line too long (handled by black)
    "S101",   # use of assert (ok in tests)
    "S104",   # possible binding to all interfaces
    "S105",   # possible hardcoded password
    "S106",   # possible hardcoded password
    "B008",   # function call in argument defaults
    "RET505", # unnecessary else after return
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ARG", "PT"]
"scripts/**/*.py" = ["T20"]

# ============================================================================
# isort - сортировка импортов
# ============================================================================
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = [
    "*/migrations/*",
    "*/venv/*",
    "*/.venv/*",
    "*/instance/*",
]
known_first_party = ["app"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

# ============================================================================
# mypy - проверка типов
# ============================================================================
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "flask_sqlalchemy.*",
    "flask_login.*",
    "flask_wtf.*",
    "wtforms.*",
    "werkzeug.*",
    "flask_migrate.*",
    "flask_limiter.*",
    "flask_caching.*",
    "flask_cors.*",
    "matplotlib.*",
    "seaborn.*",
    "plotly.*",
    "pandas.*",
    "pyotp.*",
    "qrcode.*",
]
ignore_missing_imports = true

# ============================================================================
# pytest - тестирование
# ============================================================================
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=app",
    "--cov-report=html",
    "--cov-report=term-missing",
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (with DB)",
    "slow: Slow tests",
    "smoke: Smoke tests (critical functionality)",
]

# ============================================================================
# coverage - покрытие кода
# ============================================================================
[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
    "*/migrations/*",
    "*/instance/*",
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"
